Refactor of PluginTest class
